# feat(extensions/webstorage): optimize db calls with pragmas