--- a/extensions/webstorage/lib.rs
+++ b/extensions/webstorage/lib.rs
@@ -64,6 +64,17 @@ fn get_webstorage(
       })?;
       std::fs::create_dir_all(&path.0)?;
       let conn = Connection::open(path.0.join("local_storage"))?;
+      // Enable write-ahead-logging and tweak some other stuff.
+      let initial_pragmas = "
+        -- enable write-ahead-logging mode
+        PRAGMA journal_mode=WAL;
+        PRAGMA page_size=4096;
+        PRAGMA mmap_size=6000000;
+        PRAGMA optimize;
+      ";
+
+      conn.execute_batch(initial_pragmas)?;
+      conn.set_prepared_statement_cache_capacity(128);
       conn.execute(
         "CREATE TABLE IF NOT EXISTS data (key VARCHAR UNIQUE, value VARCHAR)",
         params![],
@@ -97,7 +108,7 @@ pub fn op_webstorage_length(
 ) -> Result<u32, AnyError> {
   let conn = get_webstorage(state, persistent)?;
 
-  let mut stmt = conn.prepare("SELECT COUNT(*) FROM data")?;
+  let mut stmt = conn.prepare_cached("SELECT COUNT(*) FROM data")?;
 
   let length: u32 = stmt.query_row(params![], |row| row.get(0))?;
 
@@ -111,7 +122,8 @@ pub fn op_webstorage_key(
 ) -> Result<Option<String>, AnyError> {
   let conn = get_webstorage(state, persistent)?;
 
-  let mut stmt = conn.prepare("SELECT key FROM data LIMIT 1 OFFSET ?")?;
+  let mut stmt =
+    conn.prepare_cached("SELECT key FROM data LIMIT 1 OFFSET ?")?;
 
   let key: Option<String> = stmt
     .query_row(params![index], |row| row.get(0))
@@ -134,8 +146,8 @@ pub fn op_webstorage_set(
 ) -> Result<(), AnyError> {
   let conn = get_webstorage(state, persistent)?;
 
-  let mut stmt =
-    conn.prepare("SELECT SUM(pgsize) FROM dbstat WHERE name = 'data'")?;
+  let mut stmt = conn
+    .prepare_cached("SELECT SUM(pgsize) FROM dbstat WHERE name = 'data'")?;
   let size: u32 = stmt.query_row(params![], |row| row.get(0))?;
 
   if size >= MAX_STORAGE_BYTES {
@@ -162,7 +174,7 @@ pub fn op_webstorage_get(
 ) -> Result<Option<String>, AnyError> {
   let conn = get_webstorage(state, persistent)?;
 
-  let mut stmt = conn.prepare("SELECT value FROM data WHERE key = ?")?;
+  let mut stmt = conn.prepare_cached("SELECT value FROM data WHERE key = ?")?;
 
   let val = stmt
     .query_row(params![key_name], |row| row.get(0))
@@ -206,7 +218,7 @@ pub fn op_webstorage_iterate_keys(
 ) -> Result<Vec<String>, AnyError> {
   let conn = get_webstorage(state, persistent)?;
 
-  let mut stmt = conn.prepare("SELECT key FROM data")?;
+  let mut stmt = conn.prepare_cached("SELECT key FROM data")?;
 
   let keys = stmt
     .query_map(params![], |row| row.get::<_, String>(0))?