@littledivy 2021-07-15
> On M1 this branch is slower than main:
> 
> ```
> Benchmark #1: target/release/deno run --unstable --location=https://a.com bench_localstorage.js
>   Time (mean ± σ):      22.8 ms ±   0.9 ms    [User: 12.3 ms, System: 7.4 ms]
>   Range (min … max):    21.8 ms …  27.3 ms    101 runs
>  
> Benchmark #2: deno run --unstable --location=https://a.com bench_localstorage.js
>   Time (mean ± σ):      22.1 ms ±   1.2 ms    [User: 12.5 ms, System: 7.6 ms]
>   Range (min … max):    20.5 ms …  27.9 ms    118 runs
>  
> Summary
>   'deno run --unstable --location=https://a.com bench_localstorage.js' ran
>     1.03 ± 0.07 times faster than 'target/release/deno run --unstable --location=https://a.com bench_localstorage.js'
> ```
> 
> The test used in both our benchmarks is
> 
> ```
> var key = "name";
> var val = "x"
> 
> for(let i = 0; i < 100; i ++) localStorage.setItem(key, val);
> ```

That's almost no change since the number of runs differ and possible disk cache resulting into the tiny difference. Still quite significant improvement in my case.