@littledivy 2021-07-15
Much noticeable when running fresh without a `$DENO_DIR/location_data/`