@bartlomieju 2021-08-06
@littledivy I would be interested in landing this PR, but I'm a bit worried about problems on M1 as pointed by @crowlKats. Could you please share you benchmark code so I could also run them locally?