@evanwashere 2021-08-07
there is actually no regression on M1, @crowlKats test just didn't delete optimized sqlite database from pr test

here are benchmarks redone with more accurate measurement of improvement itself on M1

proper test:
```js
rm -rf ~/Library/Caches/deno/location_data
➜  /tmp ./deno-new run -A --location https://google.com test.js
{
  min: "38,126 ns/iter",
  max: "446,292 ns/iter",
  avg: "46,517 ns/iter",
  "50th": "45,666 ns/iter",
  "75th": "48,125 ns/iter",
  "99th": "58,167 ns/iter",
  "99.5th": "84,418 ns/iter",
  "99.9th": "202,709 ns/iter"
}
rm -rf ~/Library/Caches/deno/location_data
➜  /tmp ./deno-old run -A --location https://google.com test.js
{
  min: "203,042 ns/iter",
  max: "1,691,168 ns/iter",
  avg: "222,941 ns/iter",
  "50th": "213,251 ns/iter",
  "75th": "217,418 ns/iter",
  "99th": "337,293 ns/iter",
  "99.5th": "364,834 ns/iter",
  "99.9th": "681,251 ns/iter"
}
```

crowls test case (reusing optimized sqlite database by first run)
```js
➜  /tmp ./deno-new run -A --location https://google.com test.js
{
  min: "38,834 ns/iter",
  max: "369,750 ns/iter",
  avg: "46,337 ns/iter",
  "50th": "45,876 ns/iter",
  "75th": "48,084 ns/iter",
  "99th": "56,542 ns/iter",
  "99.5th": "63,834 ns/iter",
  "99.9th": "98,460 ns/iter"
}
➜  /tmp ./deno-old run -A --location https://google.com test.js
{
  min: "39,208 ns/iter",
  max: "382,167 ns/iter",
  avg: "47,249 ns/iter",
  "50th": "46,750 ns/iter",
  "75th": "49,043 ns/iter",
  "99th": "57,918 ns/iter",
  "99.5th": "66,085 ns/iter",
  "99.9th": "96,292 ns/iter"
}
```

source:
```js
import { sync } from 'https://gist.githubusercontent.com/evanwashere/7ee592870e46f80405b9776dcd56e1e8/raw/85f150581accc7d987b806093c32aada21463d41/bench.js';

console.log(sync(1e4, () => localStorage.setItem('a', 'aaaaa')));
```