@crowlKats 2021-08-07
I actually did clear the location_data folder back then, divy can confirm, but alright